# openapi.yaml
openapi: 3.1.0
info:
  contact:
    url: https://www.asam.net/project-detail/sovd-service-oriented-vehicle-diagnostics/
  description: Definition of ASAM SOVD API
  license:
    name: Licensing Terms for Grant of Rights to Use ASAM Products
    url: https://www.asam.net/license/
  title: Service-Oriented Vehicle Diagnostic API
  version: 1.0.0
servers:
- url: https://sovd.server/v1
...
paths:
...
  /{entity-collection}/{entity-id}/faults:
    delete:
      description: This method deletes all fault entries of an entity.
      operationId: deleteAllFaults
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - description: The scope (e.g. user-defined error memories) for which faults
          are retrieved. The capability description defines which scopes are supported.
        explode: true
        in: query
        name: scope
        required: false
        schema:
          type: string
        style: form
      responses:
        "204":
          description: The request was successful.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - fault-handling
    get:
      description: "This method provides the fault entries which are detected for\
        \ an entity, if no query parameter is used. The returned fault entries may\
        \ be filtered by their status."
      operationId: getFaults
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - description: "Specifies whether the response should include schema information\
          \ or not. Default: false"
        explode: true
        in: query
        name: include-schema
        required: false
        schema:
          type: boolean
        style: form
      - description: "Filters the available elements based on a key from the status.\
          \ The provided value is a full match against the key and value of the attribute\
          \ status in the response. For filtering by multiple status, the parameter\
          \ is repeated (0 .. *). If multiple status are provided, they are logically\
          \ OR combined. The keys supported by an entity are defined in the capability\
          \ description of the entity."
        explode: true
        in: query
        name: "status[key]"
        required: false
        schema:
          title: id
          type: string
        style: form
      - description: Filter fault entries by their severity. Any fault with a severity
          below the given number will be included in the response.
        explode: true
        in: query
        name: severity
        required: false
        schema:
          type: integer
        style: form
      - description: The scope (e.g. user-defined error memories) for which faults
          are retrieved. The capability description defines which scopes are supported.
        explode: true
        in: query
        name: scope
        required: false
        schema:
          title: id
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getFaults_200_response'
          description: Retrieve list of all faults provided by the entity
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - fault-handling
  /{entity-collection}/{entity-id}/faults/{fault-code}:
    delete:
      description: This method deletes the given fault entry from an entity.
      operationId: deleteFaultById
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - description: The scope (e.g. user-defined error memories) for which faults
          are retrieved. The capability description defines which scopes are supported.
        explode: false
        in: path
        name: fault-code
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Fault entry has been deleted successfully.
        "404":
          description: The SOVD client tried to delete a fault entry which is not
            defined for the entity.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: "The SOVD client does not fulfill the conditions for deleting\
            \ a single DTC, e.g., the DTC is safety relevant. The SOVD server should\
            \ supply SOVD REST API ASAM SOVD, API Specification Version 1.0.0 57 information\
            \ in the GenericError which condition is not met (e.g., safety conditions,\
            \ DTC not set, …)"
        "501":
          description: The request for deleting single fault entries is not supported
            by the entity.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - fault-handling
    get:
      description: This method provides the details for a fault-code.
      operationId: getFaultById
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - description: The scope (e.g. user-defined error memories) for which faults
          are retrieved. The capability description defines which scopes are supported.
        explode: false
        in: path
        name: fault-code
        required: true
        schema:
          title: id
          type: string
        style: simple
      - description: "Specifies whether the response should include schema information\
          \ or not. Default: false"
        explode: true
        in: query
        name: include-schema
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                item:
                  code: 0012E3
                  scope: Default
                  display_code: P102
                  fault_name: No signal from sensor
                  fault_translation_id: CAMERA_0012E3_tid
                  severity: 1
                  status:
                    testFailed: "1"
                    testFailedThisOperationCycle: "1"
                    pendingDTC: "0"
                    confirmedDTC: "1"
                    testNotCompletedSinceLastClear: "1"
                    testFailedSinceLastClear: "0"
                    testNotCompletedThisOperationCycle: "0"
                    warningIndicatorRequested: "0"
                    mask: 1E
                    aggregatedStatus: active
                environment_data:
                  id: env_data
                  data:
                    battery_voltage: 12.8
                    occurence_counter: 12
                    first_occurence: 2021-06-22T11:47:22Z
                    last_occurence: 2021-07-01T10:12:53Z
              schema:
                $ref: '#/components/schemas/getFaultById_200_response'
          description: The request was successful.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - fault-handling
  /ecu/engine/faults:
    delete:
      description: Delete all faults from the simulated engine ECU.
      operationId: clearEngineEcuFaults
      responses:
        "204":
          description: All faults cleared
      summary: Clear Engine Faults
      tags:
      - fault-handling
    get:
      description: "Retrieve faults from the simulated engine ECU. Supports scope\
        \ filter (active, pending, permanent)."
      operationId: getEngineEcuFaults
      parameters:
      - description: Filter faults by scope
        explode: true
        in: query
        name: scope
        required: false
        schema:
          enum:
          - active
          - pending
          - permanent
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: ./faults/faults.yaml#/components/schemas/Fault
                type: array
          description: List of engine faults
        "500":
          description: Internal server error
      summary: Get Engine Faults
      tags:
      - fault-handling
    post:
      description: Inject a simulated fault into the engine ECU (for testing purposes).
      operationId: addEngineEcuFault
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./faults/faults.yaml#/components/schemas/Fault
        required: true
      responses:
        "201":
          description: Fault created
        "400":
          description: Bad request
      summary: Add Engine Fault
      tags:
      - fault-handling
  /ecu/engine/faults/{fault_id}:
    get:
      description: Retrieve a specific fault from the engine ECU by its ID.
      operationId: getEngineEcuFaultById
      parameters:
      - description: The fault DTC code
        explode: false
        in: path
        name: fault_id
        required: true
        schema:
          title: id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: ./faults/faults.yaml#/components/schemas/Fault
          description: Fault details
        "404":
          description: Fault not found
      summary: Get Engine Fault By ID
      tags:
      - fault-handling
...
components:
  schemas:
...
    getFaultById_200_response_item:
      description: Fault description
      example:
        severity: 1
        symptom: symptom
        code: 0012E3
        fault_name: No signal from sensor
        display_code: P102
        symptom_transaction_id: symptom_transaction_id
        fault_translation_id: CAMERA_0012E3_tid
        scope: Default
        status: "{}"
      properties:
        code:
          description: Fault code in the native representation of the entity.
          example: 0012E3
          title: code
          type: string
        scope:
          description: "Defines the scope (e.g., user-defined fault memories) for\
            \ fault entries. The capability description defines which scopes are supported."
          example: Default
          title: scope
          type: string
        display_code:
          description: Display representation for the fault code.
          example: P102
          title: display_code
          type: string
        fault_name:
          description: Name / description of the fault
          example: No signal from sensor
          title: fault_name
          type: string
        fault_translation_id:
          description: Optional identifier for translating the name
          example: CAMERA_0012E3_tid
          title: fault_translation_id
          type: string
        severity:
          description: "Severity defines the impact on the availability of the vehicle.\
            \ Lower number means higher severity. For a classic ECU this resembles\
            \ the level from ODX, for HPCs the SOVD API recommends 1 = FATAL, 2 =\
            \ ERROR, 3 = WARN, and 4 = INFO."
          example: 1
          title: severity
          type: integer
        status:
          description: Detailed status information for the fault as key value pairs.
            For classic ECU this resembles the status byte of a DTC. The content of
            the keys and values is OEM-specific. The keys are defined in the capability
            description. See Notes below.
          title: status
          type: object
        symptom:
          description: Detailed symptom / failure mode information for the fault.
            For classic ECU this resembles the symptom of a DTC. The content of the
            attribute is OEM-specific.
          title: symptom
          type: string
        symptom_transaction_id:
          description: Translation identifier for the symptom name.
          items:
            anyOf:
            - {}
            - {}
            - {}
            - {}
            - items: {}
            - {}
            description: "Response parameters of the operation depending on the capability.\
              \ The type of the value is defined by the schema. Condition: If the\
              \ capability provides response parameters."
          title: symptom_transaction_id
          type: string
      required:
      - code
      - fault_name
      title: getFaultById_200_response_item
...
    getFaults_200_response_items_inner:
      description: Description of a single fault
      example:
        severity: 1
        symptom: symptom
        code: 0012E3
        fault_name: No signal from sensor
        display_code: P102
        symptom_transaction_id: symptom_transaction_id
        fault_translation_id: CAMERA_0012E3_tid
        scope: Default
        status: "{}"
      properties:
        code:
          description: Fault code in the native representation of the entity.
          example: 0012E3
          title: code
          type: string
        scope:
          description: "Defines the scope (e.g., user-defined fault memories) for\
            \ fault entries. The capability description defines which scopes are supported."
          example: Default
          title: scope
          type: string
        display_code:
          description: Display representation for the fault code.
          example: P102
          title: display_code
          type: string
        fault_name:
          description: Name / description of the fault
          example: No signal from sensor
          title: fault_name
          type: string
        fault_translation_id:
          description: Optional identifier for translating the name
          example: CAMERA_0012E3_tid
          title: fault_translation_id
          type: string
        severity:
          description: "Severity defines the impact on the availability of the vehicle.\
            \ Lower number means higher severity. For a classic ECU this resembles\
            \ the level from ODX, for HPCs the SOVD API recommends 1 = FATAL, 2 =\
            \ ERROR, 3 = WARN, and 4 = INFO."
          example: 1
          title: severity
          type: integer
        status:
          description: Detailed status information for the fault as key value pairs.
            For classic ECU this resembles the status byte of a DTC. The content of
            the keys and values is OEM-specific. The keys are defined in the capability
            description. See Notes below.
          title: status
          type: object
        symptom:
          description: Detailed symptom / failure mode information for the fault.
            For classic ECU this resembles the symptom of a DTC. The content of the
            attribute is OEM-specific.
          title: symptom
          type: string
        symptom_transaction_id:
          description: Translation identifier for the symptom name.
          items:
            anyOf:
            - {}
            - {}
            - {}
            - {}
            - items: {}
            - {}
            description: "Response parameters of the operation depending on the capability.\
              \ The type of the value is defined by the schema. Condition: If the\
              \ capability provides response parameters."
          title: symptom_transaction_id
          type: string
      required:
      - code
      - fault_name
      title: getFaults_200_response_items_inner
...
    getFaults_200_response:
      description: List of all faults of a device
      example:
        schema: "{}"
        items:
        - severity: 1
          symptom: symptom
          code: 0012E3
          fault_name: No signal from sensor
          display_code: P102
          symptom_transaction_id: symptom_transaction_id
          fault_translation_id: CAMERA_0012E3_tid
          scope: Default
          status: "{}"
        - severity: 1
          symptom: symptom
          code: 0012E3
          fault_name: No signal from sensor
          display_code: P102
          symptom_transaction_id: symptom_transaction_id
          fault_translation_id: CAMERA_0012E3_tid
          scope: Default
          status: "{}"
      properties:
        items:
          items:
            $ref: '#/components/schemas/getFaults_200_response_items_inner'
          title: items
          type: array
        schema:
          description: "Schema for describing the response. Condition: Only provided\
            \ if the query parameter include-schema is true."
          title: schema
          type: object
      required:
      - items
      title: getFaults_200_response
...
    __any_path__docs_get_default_response:
      description: It is recommended that the response body of an error response includes
        a more detailed description of the error which helps in diagnosing the issue.
      example:
        vendor_code: vendor_code
        translation_id: translation_id
        error_code: error_code
        message: message
        parameters: "{}"
      properties:
        error_code:
          description: SOVD standardized error code
          title: error_code
          type: string
        vendor_code:
          description: "Vendor specific error code Condition: If the error_code is\
            \ vendor-specific this attribute has to be filled."
          title: vendor_code
          type: string
        message:
          description: Message describing the problem in more detail.
          title: message
          type: string
        translation_id:
          description: Identifier for translating the message
          title: translation_id
          type: string
        parameters:
          description: Key-value-list which provides additional information on the
            error
          title: parameters
          type: object
      required:
      - error_code
      - message
      title: __any_path__docs_get_default_response
    __entity_collection___entity_id__data__data_id__get_200_response_errors_inner:
      description: "On some occasions an error applies only to parts of a response,\
        \ e.g., when reading a data resource. In this case the type DataError is used\
        \ which includes a path to the erroneous attribute of the response."
      example:
        path: path
        error:
          vendor_code: vendor_code
          translation_id: translation_id
          error_code: error_code
          message: message
          parameters: "{}"
      properties:
        path:
          description: A JSON Pointer describing which element of the response is
            erroneous.
          format: json-pointer
          title: path
          type: string
        error:
          $ref: '#/components/schemas/__entity_collection___entity_id__data__data_id__get_200_response_errors_inner_error'
      required:
      - path
      title: __entity_collection___entity_id__data__data_id__get_200_response_errors_inner
...
    __entity_collection___entity_id__data__data_id__get_200_response_errors_inner_error:
      description: "Describes the error more precisely. Condition: The attribute referenced\
        \ by path contains an error."
      example:
        vendor_code: vendor_code
        translation_id: translation_id
        error_code: error_code
        message: message
        parameters: "{}"
      properties:
        error_code:
          description: SOVD standardized error code
          title: error_code
          type: string
        vendor_code:
          description: "Vendor specific error code Condition: If the error_code is\
            \ vendor-specific this attribute has to be filled."
          title: vendor_code
          type: string
        message:
          description: Message describing the problem in more detail.
          title: message
          type: string
        translation_id:
          description: Identifier for translating the message
          title: translation_id
          type: string
        parameters:
          description: Key-value-list which provides additional information on the
            error
          title: parameters
          type: object
      required:
      - error_code
      - message
      title: __entity_collection___entity_id__data__data_id__get_200_response_errors_inner_error
    getFaultById_200_response:
      example:
        schema: "{}"
        item:
          severity: 1
          symptom: symptom
          code: 0012E3
          fault_name: No signal from sensor
          display_code: P102
          symptom_transaction_id: symptom_transaction_id
          fault_translation_id: CAMERA_0012E3_tid
          scope: Default
          status: "{}"
        environment_data: ""
        errors:
        - path: path
          error:
            vendor_code: vendor_code
            translation_id: translation_id
            error_code: error_code
            message: message
            parameters: "{}"
        - path: path
          error:
            vendor_code: vendor_code
            translation_id: translation_id
            error_code: error_code
            message: message
            parameters: "{}"
      properties:
        item:
          $ref: '#/components/schemas/getFaultById_200_response_item'
        environment_data:
          description: "Additional OEM-specific information describing the environment\
            \ of the system at the time of setting the fault. Condition: Only set\
            \ if the fault provides the requested information."
          title: environment_data
          example: null
        errors:
          description: "Condition: Only set, if the value environment_data represents\
            \ an error, and this attribute describes the error more precisely."
          items:
            $ref: '#/components/schemas/__entity_collection___entity_id__data__data_id__get_200_response_errors_inner'
          title: errors
          type: array
        schema:
          description: "Schema for describing the response. Condition: Only provided\
            \ if the query parameter include-schema is true."
          title: schema
          type: object
      required:
      - item
      title: getFaultById_200_response
...